More missed test.js file renames
