+++ src/components/Dropdown/Menu/Item/Item.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import Item from './';
+import Item from '.';
 
 describe('Item', () => {
   it('renders', () => {
