+++ src/components/Responsive/Hidden/Hidden.test.js
